+++ src/About/About.css
@@ -86,12 +86,6 @@
 	}
 }
 
-/* @media (width >= 1024px){
-	.about-section{
-		margin: auto;
-		max-width: 900px;
-	}
-} */
 
 /* @media (hover: hover) {
 	.about-img-box img:hover {
