+++ src/HeaderPortfolio/HeaderPortfolio.css
@@ -1,13 +1,13 @@
 @media (width >= 320px) {
 	.hero {
-		background-image: url(../src/assets/fondo1.png);
+		background-image: url(.././assets/fondo1.png);
 		background-position: center;
 		background-size: cover;
 		position: relative;
 		background-color: var(--theme-color);
 	}
 
-	.hero-overlay{
+	.hero-overlay {
 		width: 100%;
 		height: 100vh;
 		background-color: rgba(15, 18, 28, 0.8);
@@ -91,9 +91,9 @@
 	}
 }
 
-@media (width >= 960px) {
+@media (width >= 768px) {
 	.hero {
-		background-image: url(../src/assets/fondo2.png);
+		background-image: url(.././assets/fondo2.png);
 	}
 	.presentation-section {
 		flex-direction: row-reverse;
@@ -101,8 +101,8 @@
 	}
 
 	.presentation-title {
-		font-size: 3.5rem;
-		text-align: start;
+		font-size: 35px;
+		text-align: center;
 	}
 
 	.presentation-title span {
@@ -134,6 +134,13 @@
 	}
 }
 
+@media (width >= 920px) {
+	.presentation-title {
+		font-size: 3.5rem;
+		text-align: start;
+	}
+}
+
 @media (hover: hover) {
 	.download-button:hover {
 		background-color: var(--theme-color-darker);
@@ -141,4 +148,3 @@
 		transition: 0.3s all ease-in-out;
 	}
 }
-
