+++ src/HeaderPortfolio/HeaderPortfolio.jsx
@@ -1,5 +1,5 @@
 import './HeaderPortfolio.css'
-import portfolioImg from './assets/portfolio_img2_sf_1x1.png'
+import portfolioImg from '../assets/portfolio_img2_sf_1x1.png'
 
 export default function PresentationPortfolio() {
 	return (
@@ -17,7 +17,7 @@ export default function PresentationPortfolio() {
 							<div className="button-container">
 								<a
 									className="download-button"
-									href="./assets/CV_Cristhian_Salazar_J_2024.pdf"
+									href="../assets/CV_Cristhian_Salazar_J_2024.pdf"
 									download
 								>
 									Descargar CV
